Add files via upload
I have been trying for over a week to commit this document by using the gitbash command line but it keeps telling me that a) the repo media_queries doesn't exist or that my file is 3 commits ahead. I am submitting this via an upload for the grading team's consideration. but I can keep trying. hopefully, it won't count against me.